test: use fixtures from modules
