+++ packages/apps/reputation-oracle/server/src/integrations/hcaptcha/fixtures.ts
@@ -1,7 +1,10 @@
 import { faker } from '@faker-js/faker';
 import { HCaptchaConfigService } from '../../config/hcaptcha-config.service';
 
-export const mockHCaptchaConfigService: Partial<HCaptchaConfigService> = {
+export const mockHCaptchaConfigService: Omit<
+  HCaptchaConfigService,
+  'configService'
+> = {
   siteKey: faker.string.uuid(),
   apiKey: faker.string.uuid(),
   secret: `E0_${faker.string.alphanumeric()}`,
