--- packages/apps/reputation-oracle/server/test/mock-creators/web3.ts
@@ -1,9 +1,0 @@
-import { Provider } from 'ethers';
-
-export function createMockProvider(): jest.Mocked<
-  Pick<Provider, 'getFeeData'>
-> {
-  return {
-    getFeeData: jest.fn(),
-  };
-}
